wip move ecclient.Repair to ecrepairer.Repair
