+++ satellite/repair/repairer/ec.go
@@ -7,10 +7,16 @@ import (
 	"bytes"
 	"context"
 	"io"
+	"io/ioutil"
+	"sort"
+	"sync/atomic"
 	"time"
 
+	"github.com/zeebo/errs"
 	"go.uber.org/zap"
 
+	"storj.io/storj/internal/errs2"
+	"storj.io/storj/internal/sync2"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/pkcrypto"
 	"storj.io/storj/pkg/signing"
@@ -153,16 +159,6 @@ func (ec *ECRepairer) downloadAndVerifyPiece(ctx context.Context, limit *pb.Addr
 	return nil, nil
 }
 
-// Repair takes a provided segment, encodes it with the prolvided redundancy strategy,
-// and uploads the pieces in need of repair to new nodes provided by order limits.
-func (ec *ECRepairer) Repair(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, rs eestream.RedundancyStrategy, data io.Reader, expiration time.Time, timeout time.Duration, path storj.Path) (successfulNodes []*pb.Node, successfulHashes []*pb.PieceHash, err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	// TODO should be almost identical to ecclient Repair (remove ecclient Repair once implemented here)
-
-	return nil, nil, nil
-}
-
 func verifyPieceHash(ctx context.Context, limit *pb.OrderLimit, hash *pb.PieceHash, expectedHash []byte) (err error) {
 	defer mon.Task()(&ctx)(&err)
 
@@ -191,6 +187,204 @@ func verifyOrderLimitSignature(ctx context.Context, satellite signing.Signee, li
 	return nil
 }
 
+// Repair takes a provided segment, encodes it with the prolvided redundancy strategy,
+// and uploads the pieces in need of repair to new nodes provided by order limits.
+func (ec *ECRepairer) Repair(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, rs eestream.RedundancyStrategy, data io.Reader, expiration time.Time, timeout time.Duration, path storj.Path) (successfulNodes []*pb.Node, successfulHashes []*pb.PieceHash, err error) {
+	defer mon.Task()(&ctx)(&err)
+
+	defer mon.Task()(&ctx)(&err)
+
+	pieceCount := len(limits)
+	if pieceCount != rs.TotalCount() {
+		return nil, nil, Error.New("size of limits slice (%d) does not match total count (%d) of erasure scheme", pieceCount, rs.TotalCount())
+	}
+
+	if !unique(limits) {
+		return nil, nil, Error.New("duplicated nodes are not allowed")
+	}
+
+	padded := eestream.PadReader(ioutil.NopCloser(data), rs.StripeSize())
+	readers, err := eestream.EncodeReader(ctx, ec.log, padded, rs)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	type info struct {
+		i    int
+		err  error
+		hash *pb.PieceHash
+	}
+	infos := make(chan info, pieceCount)
+
+	psCtx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	for i, addressedLimit := range limits {
+		go func(i int, addressedLimit *pb.AddressedOrderLimit) {
+			hash, err := ec.putPiece(psCtx, ctx, addressedLimit, privateKey, readers[i], expiration)
+			infos <- info{i: i, err: err, hash: hash}
+		}(i, addressedLimit)
+	}
+	ec.log.Info("Starting a timer for repair so that the number of pieces will be closer to the success threshold",
+		zap.Duration("Timer", timeout),
+		zap.String("Path", path),
+		zap.Int("Node Count", nonNilCount(limits)),
+		zap.Int("Optimal Threshold", rs.OptimalThreshold()),
+	)
+
+	var successfulCount, failureCount, cancellationCount int32
+	timer := time.AfterFunc(timeout, func() {
+		if ctx.Err() != context.Canceled {
+			ec.log.Info("Timer expired. Canceling the long tail...",
+				zap.String("Path", path),
+				zap.Int32("Successfully repaired", atomic.LoadInt32(&successfulCount)),
+			)
+			cancel()
+		}
+	})
+
+	successfulNodes = make([]*pb.Node, pieceCount)
+	successfulHashes = make([]*pb.PieceHash, pieceCount)
+
+	for range limits {
+		info := <-infos
+
+		if limits[info.i] == nil {
+			continue
+		}
+
+		if info.err != nil {
+			if !errs2.IsCanceled(info.err) {
+				failureCount++
+			} else {
+				cancellationCount++
+			}
+			ec.log.Debug("Repair to storage node failed",
+				zap.String("Path", path),
+				zap.String("NodeID", limits[info.i].GetLimit().StorageNodeId.String()),
+				zap.Error(info.err),
+			)
+			continue
+		}
+
+		successfulNodes[info.i] = &pb.Node{
+			Id:      limits[info.i].GetLimit().StorageNodeId,
+			Address: limits[info.i].GetStorageNodeAddress(),
+		}
+		successfulHashes[info.i] = info.hash
+		successfulCount++
+	}
+
+	// Ensure timer is stopped
+	_ = timer.Stop()
+
+	// TODO: clean up the partially uploaded segment's pieces
+	defer func() {
+		select {
+		case <-ctx.Done():
+			err = Error.New("repair cancelled")
+			// ec.Delete(context.Background(), nodes, pieceID, pba.SatelliteId), //TODO
+		default:
+		}
+	}()
+
+	if successfulCount == 0 {
+		return nil, nil, Error.New("repair %v to all nodes failed", path)
+	}
+
+	ec.log.Info("Successfully repaired",
+		zap.String("Path", path),
+		zap.Int32("Success Count", atomic.LoadInt32(&successfulCount)),
+	)
+
+	mon.IntVal("repair_segment_pieces_total").Observe(int64(pieceCount))
+	mon.IntVal("repair_segment_pieces_successful").Observe(int64(successfulCount))
+	mon.IntVal("repair_segment_pieces_failed").Observe(int64(failureCount))
+	mon.IntVal("repair_segment_pieces_canceled").Observe(int64(cancellationCount))
+
+	return successfulNodes, successfulHashes, nil
+}
+
+// TODO limit duplicate code with ecclient
+
+// copied from ecclient
+func (ec *ECRepairer) putPiece(ctx, parent context.Context, limit *pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, data io.ReadCloser, expiration time.Time) (hash *pb.PieceHash, err error) {
+	nodeName := "nil"
+	if limit != nil {
+		nodeName = limit.GetLimit().StorageNodeId.String()[0:8]
+	}
+	defer mon.Task()(&ctx, "node: "+nodeName)(&err)
+	defer func() { err = errs.Combine(err, data.Close()) }()
+
+	if limit == nil {
+		_, _ = io.Copy(ioutil.Discard, data)
+		return nil, nil
+	}
+
+	storageNodeID := limit.GetLimit().StorageNodeId
+	pieceID := limit.GetLimit().PieceId
+	ps, err := ec.dialPiecestore(ctx, &pb.Node{
+		Id:      storageNodeID,
+		Address: limit.GetStorageNodeAddress(),
+	})
+	if err != nil {
+		ec.log.Debug("Failed dialing for putting piece to node",
+			zap.String("PieceID", pieceID.String()),
+			zap.String("NodeID", storageNodeID.String()),
+			zap.Error(err),
+		)
+		return nil, err
+	}
+	defer func() { err = errs.Combine(err, ps.Close()) }()
+
+	upload, err := ps.Upload(ctx, limit.GetLimit(), privateKey)
+	if err != nil {
+		ec.log.Debug("Failed requesting upload of pieces to node",
+			zap.String("PieceID", pieceID.String()),
+			zap.String("NodeID", storageNodeID.String()),
+			zap.Error(err),
+		)
+		return nil, err
+	}
+	defer func() {
+		if ctx.Err() != nil || err != nil {
+			hash = nil
+			err = errs.Combine(err, upload.Cancel(ctx))
+			return
+		}
+		h, closeErr := upload.Commit(ctx)
+		hash = h
+		err = errs.Combine(err, closeErr)
+	}()
+
+	_, err = sync2.Copy(ctx, upload, data)
+	// Canceled context means the piece upload was interrupted by user or due
+	// to slow connection. No error logging for this case.
+	if ctx.Err() == context.Canceled {
+		if parent.Err() == context.Canceled {
+			ec.log.Info("Upload to node canceled by user", zap.String("NodeID", storageNodeID.String()))
+		} else {
+			ec.log.Debug("Node cut from upload due to slow connection", zap.String("NodeID", storageNodeID.String()))
+		}
+		err = context.Canceled
+	} else if err != nil {
+		nodeAddress := "nil"
+		if limit.GetStorageNodeAddress() != nil {
+			nodeAddress = limit.GetStorageNodeAddress().GetAddress()
+		}
+
+		ec.log.Debug("Failed uploading piece to node",
+			zap.String("PieceID", pieceID.String()),
+			zap.String("NodeID", storageNodeID.String()),
+			zap.String("Node Address", nodeAddress),
+			zap.Error(err),
+		)
+	}
+
+	return hash, err
+}
+
+// copied from ecclient
 func calcPadded(size int64, blockSize int) int64 {
 	mod := size % int64(blockSize)
 	if mod == 0 {
@@ -199,6 +393,7 @@ func calcPadded(size int64, blockSize int) int64 {
 	return size + int64(blockSize) - mod
 }
 
+// copied from ecclient
 func nonNilCount(limits []*pb.AddressedOrderLimit) int {
 	total := 0
 	for _, limit := range limits {
@@ -207,4 +402,28 @@ func nonNilCount(limits []*pb.AddressedOrderLimit) int {
 		}
 	}
 	return total
+}
+
+// copied from ecclient
+func unique(limits []*pb.AddressedOrderLimit) bool {
+	if len(limits) < 2 {
+		return true
+	}
+	ids := make(storj.NodeIDList, len(limits))
+	for i, addressedLimit := range limits {
+		if addressedLimit != nil {
+			ids[i] = addressedLimit.GetLimit().StorageNodeId
+		}
+	}
+
+	// sort the ids and check for identical neighbors
+	sort.Sort(ids)
+	// sort.Slice(ids, func(i, k int) bool { return ids[i].Less(ids[k]) })
+	for i := 1; i < len(ids); i++ {
+		if ids[i] != (storj.NodeID{}) && ids[i] == ids[i-1] {
+			return false
+		}
+	}
+
+	return true
 }
