+++ uplink/ecclient/client.go
@@ -31,7 +31,6 @@ var mon = monkit.Package()
 // Client defines an interface for storing erasure coded data to piece store nodes
 type Client interface {
 	Put(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, rs eestream.RedundancyStrategy, data io.Reader, expiration time.Time) (successfulNodes []*pb.Node, successfulHashes []*pb.PieceHash, err error)
-	Repair(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, rs eestream.RedundancyStrategy, data io.Reader, expiration time.Time, timeout time.Duration, path storj.Path) (successfulNodes []*pb.Node, successfulHashes []*pb.PieceHash, err error)
 	Get(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, es eestream.ErasureScheme, size int64) (ranger.Ranger, error)
 	Delete(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey) error
 	WithForceErrorDetection(force bool) Client
@@ -175,120 +174,6 @@ func (ec *ecClient) Put(ctx context.Context, limits []*pb.AddressedOrderLimit, p
 	if successes < rs.OptimalThreshold() {
 		return nil, nil, Error.New("successful puts (%d) less than success threshold (%d)", successes, rs.OptimalThreshold())
 	}
-
-	return successfulNodes, successfulHashes, nil
-}
-
-func (ec *ecClient) Repair(ctx context.Context, limits []*pb.AddressedOrderLimit, privateKey storj.PiecePrivateKey, rs eestream.RedundancyStrategy, data io.Reader, expiration time.Time, timeout time.Duration, path storj.Path) (successfulNodes []*pb.Node, successfulHashes []*pb.PieceHash, err error) {
-	defer mon.Task()(&ctx)(&err)
-
-	pieceCount := len(limits)
-	if pieceCount != rs.TotalCount() {
-		return nil, nil, Error.New("size of limits slice (%d) does not match total count (%d) of erasure scheme", pieceCount, rs.TotalCount())
-	}
-
-	if !unique(limits) {
-		return nil, nil, Error.New("duplicated nodes are not allowed")
-	}
-
-	padded := eestream.PadReader(ioutil.NopCloser(data), rs.StripeSize())
-	readers, err := eestream.EncodeReader(ctx, ec.log, padded, rs)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	type info struct {
-		i    int
-		err  error
-		hash *pb.PieceHash
-	}
-	infos := make(chan info, pieceCount)
-
-	psCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	for i, addressedLimit := range limits {
-		go func(i int, addressedLimit *pb.AddressedOrderLimit) {
-			hash, err := ec.putPiece(psCtx, ctx, addressedLimit, privateKey, readers[i], expiration)
-			infos <- info{i: i, err: err, hash: hash}
-		}(i, addressedLimit)
-	}
-	ec.log.Info("Starting a timer for repair so that the number of pieces will be closer to the success threshold",
-		zap.Duration("Timer", timeout),
-		zap.String("Path", path),
-		zap.Int("Node Count", nonNilCount(limits)),
-		zap.Int("Optimal Threshold", rs.OptimalThreshold()),
-	)
-
-	var successfulCount, failureCount, cancellationCount int32
-	timer := time.AfterFunc(timeout, func() {
-		if ctx.Err() != context.Canceled {
-			ec.log.Info("Timer expired. Canceling the long tail...",
-				zap.String("Path", path),
-				zap.Int32("Successfully repaired", atomic.LoadInt32(&successfulCount)),
-			)
-			cancel()
-		}
-	})
-
-	successfulNodes = make([]*pb.Node, pieceCount)
-	successfulHashes = make([]*pb.PieceHash, pieceCount)
-
-	for range limits {
-		info := <-infos
-
-		if limits[info.i] == nil {
-			continue
-		}
-
-		if info.err != nil {
-			if !errs2.IsCanceled(info.err) {
-				failureCount++
-			} else {
-				cancellationCount++
-			}
-			ec.log.Debug("Repair to storage node failed",
-				zap.String("Path", path),
-				zap.String("NodeID", limits[info.i].GetLimit().StorageNodeId.String()),
-				zap.Error(info.err),
-			)
-			continue
-		}
-
-		successfulNodes[info.i] = &pb.Node{
-			Id:      limits[info.i].GetLimit().StorageNodeId,
-			Address: limits[info.i].GetStorageNodeAddress(),
-		}
-		successfulHashes[info.i] = info.hash
-		successfulCount++
-	}
-
-	// Ensure timer is stopped
-	_ = timer.Stop()
-
-	// TODO: clean up the partially uploaded segment's pieces
-	defer func() {
-		select {
-		case <-ctx.Done():
-			err = Error.New("repair cancelled")
-			// ec.Delete(context.Background(), nodes, pieceID, pba.SatelliteId), //TODO
-		default:
-		}
-	}()
-
-	if successfulCount == 0 {
-		return nil, nil, Error.New("repair %v to all nodes failed", path)
-	}
-
-	ec.log.Info("Successfully repaired",
-		zap.String("Path", path),
-		zap.Int32("Success Count", atomic.LoadInt32(&successfulCount)),
-	)
-
-	mon.IntVal("repair_segment_pieces_total").Observe(int64(pieceCount))
-	mon.IntVal("repair_segment_pieces_successful").Observe(int64(successfulCount))
-	mon.IntVal("repair_segment_pieces_failed").Observe(int64(failureCount))
-	mon.IntVal("repair_segment_pieces_canceled").Observe(int64(cancellationCount))
 
 	return successfulNodes, successfulHashes, nil
 }
